+++ src/styles/themes/next-gen/variants/environmentBreadcrumb.ts
@@ -1,0 +1,10 @@
+const button = {
+  current: {
+    color: 'font.base',
+    fontWeight: 2,
+  },
+};
+
+export default {
+  button,
+};