+++ CHANGELOG.md
@@ -1,3 +1,33 @@
+## 2.194.0-alpha.0 (2026-04-30)
+
+### 🚀 Features
+
+- [UXE-8273] add onyx overrides and add to onyx wrapper
+
+### ❤️  Thank You
+
+- Andrew Keohane
+
+## 2.193.2-alpha.0 (2026-04-30)
+
+### 🩹 Fixes
+
+- [UIP-8186] update Country Picker recipe
+
+### ❤️  Thank You
+
+- Kiran Gupta
+
+## 2.193.1-alpha.0 (2026-04-29)
+
+### 🩹 Fixes
+
+- [UXE-8166] EnvironmentBreadcrumb-design-fixes
+
+### ❤️  Thank You
+
+- Andrew Keohane
+
 ## 2.193.0 (2026-04-28)
 
 This was a version bump only for astro to align it with other projects, there were no code changes.